renaming, singletones
