+++ lib/data/interactors/settings_interactor.dart
@@ -1,0 +1,9 @@
+// Напишите SettingsInteractor в котором происходит логика смены темной/светлой темы
+// Подключите интерактор к экрану настроек
+
+class SettingsInteractor {
+  bool isDartThemeOn = false;
+  void toggleTheme() {
+    isDartThemeOn = !isDartThemeOn;
+  }
+}
