+++ lib/data/repositories/search_repository.dart
@@ -1,0 +1,3 @@
+class SearchRepository {
+  
+}
