+++ lib/data/repositories/settings_repository.dart
@@ -1,0 +1,3 @@
+class SettingsRepository {
+  
+}
