+++ lib/ui/screens/main_screen/main_screen_model.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 import 'package:places/ui/screens/sight_details_screen/sight_details_screen.dart';
 
-import 'package:places/data/repositories/mocks.dart';
-import 'package:places/data/models/sight.dart';
+import 'package:places/data/repositories/place_repository.dart';
+import 'package:places/data/models/place.dart';
 
 
 class MainScreenModel with ChangeNotifier {
