+++ lib/ui/screens/sight_details_screen/sight_details_model.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/data/interactors/core.dart';
-import 'package:places/data/models/sight.dart';
+import 'package:places/data/interactors/_core.dart';
+import 'package:places/data/models/place.dart';
 
 class SightDetailsModel with ChangeNotifier {
   var core = CoreSingleton();
