+++ lib/ui/widgets/sight_card.dart
@@ -4,7 +4,7 @@ import 'package:places/enums.dart';
 import 'package:places/ui/widgets/my_image_widget.dart';
 import 'package:places/ui/widgets/sight_card_icons.dart';
 
-import 'package:places/data/models/sight.dart';
+import 'package:places/data/models/place.dart';
 
 class SightCard extends StatelessWidget {
   const SightCard({
